An "Email Reader"
Is a quick and dirty way to see what information applications are trying to send to your desktop's mail client, without having to set up a mail client! Just make this the default and pass the string to it as the first argument. It will open in Pluma, though you'll want to change that to your own text editor.